+++ keystore.cpp
@@ -1,0 +1,35 @@
+//
+// Created by dhruva on 1/10/26.
+//
+
+#include "keystore.h"
+
+#include <format>
+#include <iostream>
+
+keystore::keystore(WaLogger &logger) : logger(logger) {
+    logger.addLogEntry(OpType::Internal, "keystore", "created");
+}
+
+keystore::~keystore() = default;
+
+int keystore::put(std::string key, std::string value) {
+    logger.addLogEntry(OpType::Put, key, value);
+    store.insert({key, value});
+
+    const auto display_string = std::format("Inserted {}:{} in store", key, value);
+    std::cout << display_string;
+    return 0;
+}
+
+std::string keystore::get(const std::string &key) {
+    auto value = store.at(key);
+    logger.addLogEntry(OpType::Get, key, value);
+    return value;
+}
+
+int keystore::del(const std::string& key) {
+    auto value = store.at(key);
+    logger.addLogEntry(OpType::Delete, key, value);
+    return store.erase(key);
+}